+++ src/presentation/helpers/validators/index.ts
@@ -1,0 +1,4 @@
+export * from './compare-fields-validation'
+export * from './email-validation'
+export * from './required-field-validation'
+export * from './validation-composite'
